% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{CR}
\alias{CR}
\alias{CR.matrix}
\title{Cressie-Read Power-Divergence Statistic}
\usage{
CR(x, lambda = 2/3)

\method{CR}{matrix}(x, lambda = 2/3)
}
\arguments{
\item{x}{A matrix of cross-tabulated counts.}

\item{lambda}{The model parameter of the Cressie-Read power-divergence statistic. Default is the recommended value of 2/3.}
}
\value{
The Cressie-Read power-divergence statistic.
}
\description{
For a matrix of cross-tabulated counts of observations, computes the Cressie-Read power-divergence statistic according to the selection of a parameter \eqn{\lambda} \insertCite{cressie_multinomial_1984,read_goodness--fit_1988}{arkhaia}.
}
\examples{
x1 <- c(2, 0, 10, 11, 5)
x2 <- c(1, 1, 17, 23, 3)
x3 <- c(0, 0, 2, 81, 11)

x <- matrix(c(x1, x2, x3), ncol = 3)

CR(x)
CR(x, lambda = 1)

}
\references{
\insertAllCited{}
}
