% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{effect_sizes}
\alias{effect_sizes}
\title{Effect Sizes for Related vs. Unrelated Contexts}
\usage{
effect_sizes(x, related = NULL, unrelated = NULL, direction = "UW")

effect_sizes(x, related = NULL, unrelated = NULL, direction = "UW")
}
\arguments{
\item{x}{A data frame in which the first column is labeled \code{Context} and the second column is labeled \code{Type}, with each row containing a type present in that context.}

\item{related}{The related pairs of contexts.}

\item{unrelated}{The unrelated pairs of contexts.}

\item{direction}{Whether the related or unrelated effect size should come first. Default is \code{"UW"}; alternative is \code{"WU"}.}
}
\value{
A list of results.
}
\description{
Selecting all pairs of columns (contexts) of a contingency table, computes the trace of a 2 x 2 contingency table that indicates the number of matching types (rows) that are present.
}
