% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{homogeneity}
\alias{homogeneity}
\alias{homogeneity.effect_sizes}
\title{Homogeneity of Related Assemblages via Effect Size}
\usage{
homogeneity(x, related = NULL, unrelated = NULL, direction = "UW")

\method{homogeneity}{effect_sizes}(x, related = NULL, unrelated = NULL, direction = "UW")
}
\arguments{
\item{x}{An effect_sizes object as returned by \code{\link[arkhaia]{VB_pair}}.}

\item{related}{The related pairs of contexts as a two-column matrix or data frame (contexts between which one anticipates a meaningful relationship). Names must match \code{colnames(x)}.}

\item{unrelated}{The unrelated pairs of contexts as a two-column matrix or data frame (contexts between which one does not anticipate a meaningful relationship). Names must match \code{colnames(x)}. May be left \code{NULL}, in which event all pairs not expressed in \code{related} are created}

\item{direction}{Whether the related or unrelated effect size should come first. Default is \code{"UW"}; alternative is \code{"WU"}.}
}
\value{
A list of results:
 * \code{U} : The effect sizes between unrelated pairs of contexts.
 * \code{W} : The effect sizes between related pairs of contexts.
 * \code{Q} : The quantile indicating the proportion of related contexts more homogenous than unrelated contexts (if direction is \code{"UW"}); less homogenous if direction is set to \code{"UW"}.
 * \code{D} : The distribution of differences, \eqn{D_{ij} = U_j - W_i}, if direction is set to \code{"UW"}. The proportion of \eqn{D > 0} is equivalent to the mean of \code{Q}.
}
\description{
Given a contingency table of counts, with contexts along the columns and types along rows, as well as identified pairs of contexts ass
}
\examples{
x1 <- c(2, 0, 10, 11, 5)
x2 <- c(1, 1, 17, 23, 3)
x3 <- c(0, 0, 2, 81, 11)
x4 <- c(3, 18, 9, 0, 23)
x <- matrix(c(x1, x2, x3, x4), ncol = 4)
colnames(x) <- c("surface1", "subsurface1", "surface2", "subsurface2")
rownames(x) <- LETTERS[1:5]

# related pairs
W_contexts <- matrix(c("surface1", "surface2", "subsurface1", "subsurface2"), ncol = 2)

# unrelated pairs (will be automatically created if left NULL)
U_contexts <- matrix(c("surface1", "surface1", "surface2", "subsurface1",
                       "surface2","subsurface2", "surface1", "subsurface2"), ncol = 2)

effect_sizes <- VB_pair(x)
homogeneity(effect_sizes, related = W_contexts, unrelated = U_contexts) 
homogeneity(effect_sizes, related = W_contexts) 


}
