% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{homogeneity_LOO}
\alias{homogeneity_LOO}
\alias{homogeneity_LOO.matrix}
\alias{homogeneity_LOO.data.frame}
\alias{homogeneity_LOO.table}
\alias{homogeneity_LOO.xtabs}
\title{Leave-One-Out Routine for Homogeneity of Related Assemblages}
\usage{
homogeneity_LOO(
  x,
  related = NULL,
  unrelated = NULL,
  direction = "UW",
  method = "CR",
  lambda = 2/3,
  simulate = TRUE
)

\method{homogeneity_LOO}{matrix}(
  x,
  related = NULL,
  unrelated = NULL,
  direction = "UW",
  method = "CR",
  lambda = 2/3,
  simulate = TRUE
)

\method{homogeneity_LOO}{data.frame}(
  x,
  related = NULL,
  unrelated = NULL,
  direction = "UW",
  method = "CR",
  lambda = 2/3,
  simulate = TRUE
)

\method{homogeneity_LOO}{table}(
  x,
  related = NULL,
  unrelated = NULL,
  direction = "UW",
  method = "CR",
  lambda = 2/3,
  simulate = TRUE
)

\method{homogeneity_LOO}{xtabs}(
  x,
  related = NULL,
  unrelated = NULL,
  direction = "UW",
  method = "CR",
  lambda = 2/3,
  simulate = TRUE
)
}
\arguments{
\item{x}{A data frame or matrix representing a contingency table of counts, with contexts along the columns and types along rows.}

\item{related}{The related pairs of contexts as a two-column matrix or data frame (contexts between which one anticipates a meaningful relationship). Names must match \code{colnames(x)}.}

\item{unrelated}{The unrelated pairs of contexts as a two-column matrix or data frame (contexts between which one does not anticipate a meaningful relationship). Names must match \code{colnames(x)}.}

\item{direction}{Whether the related or unrelated effect size should come first. Default is \code{"UW"}; alternative is \code{"WU"}.}

\item{method}{Method for estimating chi squared. Default is \code{CR}, the Cressie-Read method (see \code{\link[arkhaia]{CR}})). Other options include \code{Pearson}, which uses R's base \code{\link[stats]{chisq.test}}) function, and \code{G}, which computes the likelihood-ratio \eqn{G} statistic, which is the case for the Cressie-Read statistic as \eqn{\lambda} approaches 0.}

\item{lambda}{Only applied to the method \code{CR}. Default is the recommended value of 2/3.}

\item{simulate}{Only applied ot the method \code{Pearson}.  Whether to simulate \eqn{p} values in the \code{\link[stats]{chisq.test}}) function. Default is \code{TRUE}.}
}
\value{
A list containing:
  * \code{EQ} : The mean quantile expressing the degree of homogeneity among related contexts as computed by the \code{\link[arkhaia]{homogeneity}} function.
  * \code{EQ_T_mean} : The mean quantile upon iterating over the omission of each type (of the LOO samples). 
  * \code{EQ_T_var} : The variance of the LOO type samples.
  * \code{EQ_C_mean} : The mean quantile upon iterating over the omission of each context.
  * \code{EQ_C_var} : The variance of the LOO context samples.
}
\description{
To evaluate the stability of estimates of effect sizes between archaeological contexts in light of the inclusion or exclusion of types and contexts, this routine computes the homogeneity of related assemblages iteratively in two ways, first by leaving out a type on each iteration and second by leaving out a context on each iteration (see \code{\link[arkhaia]{homogeneity}} and \code{\link[arkhaia]{VB}}).
}
\examples{
x1 <- c(2, 0, 10, 11, 5)
x2 <- c(1, 1, 17, 23, 3)
x3 <- c(0, 0, 2, 81, 11)
x4 <- c(3, 18, 9, 0, 23)
x <- matrix(c(x1, x2, x3, x4), ncol = 4)
colnames(x) <- c("surface1", "subsurface1", "surface2", "subsurface2")
rownames(x) <- LETTERS[1:5]

# related pairs
W_contexts <- matrix(c("surface1", "surface2", "subsurface1", "subsurface2"), ncol = 2)

# unrelated pairs (will be automatically created if left NULL)
U_contexts <- matrix(c("surface1", "surface1", "surface2", "subsurface1",
                       "surface2","subsurface2", "surface1", "subsurface2"), ncol = 2)

homogeneity_LOO(x, related = W_contexts, unrelated = U_contexts) 

}
