% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{log_OR}
\alias{log_OR}
\alias{log_OR.matrix}
\title{Trace Coefficient}
\usage{
log_OR(x)

\method{log_OR}{matrix}(x)
}
\arguments{
\item{x}{A data frame in which the first column is labeled \code{Context} and the second column is labeled \code{Type}, with each row containing a type present in that context.}
}
\value{
A three-dimensional array of the pairwise context-by-context effect sizes given the ommission of a type, for all types given in the input data frame.
}
\description{
Computes the trace of the
}
