% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{log_OR_pair}
\alias{log_OR_pair}
\alias{log_OR_pair.matrix}
\title{Pairwise Log Odds Ratios with Haldane-Anscombe's Correction}
\usage{
log_OR_pair(x)

\method{log_OR_pair}{matrix}(x)
}
\arguments{
\item{x}{A data frame in which the first column is labeled \code{Context} and the second column is labeled \code{Type}, with each row containing a type present in that context.}
}
\value{
A matrix of the pairwise log odds ratios giving the overlap in types between contexts.
}
\description{
Selecting all pairs of columns (contexts) of a contingency table, computes the log odds ratio of a 2 x 2 contingency table that indicates the number of matching types (rows) that are present. The addition of 0.5 to cells of the contingency table is performed  \insertCite{anscombe_estimating_1956,haldane_estimation_1956}{arkhaia}.
}
