% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{LSSA}
\alias{LSSA}
\alias{LSSA.matrix}
\alias{LSSA.data.frame}
\title{Least Squares Spectral Analysis (LSSA)}
\usage{
LSSA(
  x,
  freqs = seq(0.001, 0.5, by = 5e-04),
  intercept = TRUE,
  type = "frequency"
)

\method{LSSA}{matrix}(
  x,
  freqs = seq(0.001, 0.5, by = 5e-04),
  intercept = TRUE,
  type = "frequency"
)

\method{LSSA}{data.frame}(
  x,
  freqs = seq(0.001, 0.5, by = 5e-04),
  intercept = TRUE,
  type = "frequency"
)
}
\arguments{
\item{x}{A data frame of two columns, the first containing time indices and the second containing values.}

\item{freqs}{A vector of frequencies to evaluate. By default a grid from 0.001 to 0.5 is tested at an interval of 0.005.}

\item{intercept}{Whether to include the intercept. Default is \code{TRUE}.}

\item{type}{Type of output. Can be either \code{"frequency"} (the default) or \code{"period"}.}
}
\value{
The a data frame containing the power and residual sum of squares for each frequency, as well as coefficients.
}
\description{
Peforms a simple least squares fitting to time indexed data of the form \eqn{x(t) = \beta_{0} + \sum_{i =1}^n \beta_{1i} \cos (2 \pi f t) + \beta_{2i} \sin(2 \pi f t)}, using a range of potential frequencies \insertCite{vanicek_approximate_1969,vanicek_further_1971}{arkhaia}. Intercept may be ommitted.
}
\references{
\insertAllCited{}
}
