% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{LSSA_LFI}
\alias{LSSA_LFI}
\alias{LSSA_LFI.matrix}
\alias{LSSA_LFI.data.frame}
\title{Least Squares Spectral Analysis via Lowest Frequency Iteration (LSSA-LFI)}
\usage{
LSSA_LFI(x, n_iter = 1, intercept = TRUE, AIC = FALSE)

\method{LSSA_LFI}{matrix}(x, n_iter = 1, intercept = TRUE, AIC = FALSE)

\method{LSSA_LFI}{data.frame}(x, n_iter = 1, intercept = TRUE, AIC = FALSE)
}
\arguments{
\item{x}{A data frame of two columns, with the first column containing time indices and the second containing values.}

\item{n_iter}{The number of iterations to run. Default is 1.}

\item{intercept}{Whether to include the intercept. Default is \code{TRUE}.}

\item{AIC}{If \code{TRUE}, only the result that has yeilded the lowest AIC (Aikake Information Criterion) is given. Default is \code{FALSE}.}
}
\value{
A list containing:
  * A list of the coefficients for each iteration (the intercept is included in the first iteration).
  * A vector of the frequencies.
  * The residual sum of squares (RSS) after each iteration (decreasing).
  * The AIC upon each iteration. (If the paraemter AIC is \code{TRUE}, this will stop at the lowest AIC value produced by the frequencies tested).
}
\description{
Peforms a simple least squares fitting to time indexed data of the form y = SIGMA B cos(2 pi f t) + C sin(2 pi f t) + D, using an input of frequencies; the lowest frequency peak (not the highest power frequency) is chosen for regression, up to a chosen number of iterations. Intercept may be ommitted. The lowest frequency is equivalent to the longest period.
}
