% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{LSSA_LFI_epoch}
\alias{LSSA_LFI_epoch}
\title{Period-Variable Pairwise Selection of Linearlly Dependent LSSA-LFI Models}
\usage{
LSSA_LFI_epoch(
  x,
  pair = NULL,
  n_iter = 1,
  intercept = TRUE,
  t_range = NULL,
  h_range = NULL,
  t_grid = 1,
  h_grid = 1
)

LSSA_LFI_epoch(
  x,
  pair = NULL,
  n_iter = 1,
  intercept = TRUE,
  t_range = NULL,
  h_range = NULL,
  t_grid = 1,
  h_grid = 1
)
}
\arguments{
\item{x}{A list of data frames.}

\item{pair}{The pair of commodities to evaluate in the list \code{x}, either names or indices.}

\item{n_iter}{The number of iterations to run. Default is 1.}

\item{intercept}{Whether to include the intercept in the least squares spectral analysis via lowest frequency iteration (LSSA-LFI). Default is \code{TRUE}.}

\item{t_range}{The range of the time period. Default are the minimum and maximum dates spanned by the data in \code{x}.}

\item{h_range}{The range of the potential windows of \eqn{h}. Default is from 0 to entire span of \code{t_range}.}

\item{t_grid}{The length of interval along which to sample \eqn{t}. Default is 1.}

\item{h_grid}{The length of interval along which to sample \eqn{h}. Default is 1.}
}
\value{
A matrix giving the probability of linear dependence, predicated upon time \eqn{t} and window length \eqn{h}.
}
\description{
Evaluate pairwise linear dependence between two time series using a LSSA-LFI valdiated model selection (see \code{\link[arkhaia]{LSSA_LFI_validated}})), with variable length time period.
}
