% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{LSSA_LFI_model}
\alias{LSSA_LFI_model}
\alias{LSSA_LFI_model.matrix}
\alias{LSSA_LFI_model.data.frame}
\title{LSSA-LFI Model}
\usage{
LSSA_LFI_model(x, t_ = NULL, n_iter = 1, intercept = TRUE, label = "model")

\method{LSSA_LFI_model}{matrix}(x, t_ = NULL, n_iter = 1, intercept = TRUE, label = "model")

\method{LSSA_LFI_model}{data.frame}(x, t_ = NULL, n_iter = 1, intercept = TRUE, label = "model")
}
\arguments{
\item{x}{A data frame where time indices are in the first column and values are in the second.}

\item{t_}{A vector giving samples range of \eqn{t} for computing \eqn{f(t)}. Default is from the minimum to maximum time index sampled at 0.01 intervals.}

\item{n_iter}{The number of iterations to run. Default is 1.}

\item{intercept}{Whether to include the intercept in the least squares spectral analysis via lowest frequency iteration (LSSA-LFI). Default is \code{TRUE}.}

\item{label}{Default is \code{"model"}.}
}
\value{
A data frame containing \eqn{t, f(t)}.
}
\description{
Generates a data frame of values \eqn{f(t)} of the model generated by LSSA-LFI (see \code{\link[arkhaia]{LSSA_LFI}})).
}
