% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{LSSA_LFI_pairwise}
\alias{LSSA_LFI_pairwise}
\alias{LSSA_LFI_pairwise.list}
\title{Pairwise Selection of Linearlly Dependent LSSA-LFI Models}
\usage{
LSSA_LFI_pairwise(x, n_iter = 1, intercept = TRUE)

\method{LSSA_LFI_pairwise}{list}(x, n_iter = 1, intercept = TRUE)
}
\arguments{
\item{x}{A list of data frames.}

\item{n_iter}{The number of iterations to run. Default is 1.}

\item{intercept}{Whether to include the intercept in the least squares spectral analysis via lowest frequency iteration (LSSA-LFI). Default is \code{TRUE}.}
}
\value{
An upper-triangular matrix, containing the probability of linear dependence between series.
}
\description{
Evaluate pairwise linear dependence between observations using a LSSA-LFI valdiated model selection (see \code{\link[arkhaia]{LSSA_LFI_validated}})).
}
