% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{LSSA_LFI_validated}
\alias{LSSA_LFI_validated}
\alias{LSSA_LFI_validated.list}
\title{Validated Linear Dependence via LSSA-LFI}
\usage{
LSSA_LFI_validated(x, pair = NULL, n_iter = 1, intercept = TRUE)

\method{LSSA_LFI_validated}{list}(x, pair = NULL, n_iter = 1, intercept = TRUE)
}
\arguments{
\item{x}{A list of data frames.}

\item{pair}{The pair of commodities to evaluate in the list \code{x}, either names or indices.}

\item{n_iter}{The number of iterations to run. Default is 1.}

\item{intercept}{Whether to include the intercept in the least squares spectral analysis via lowest frequency iteration (LSSA-LFI). Default is \code{TRUE}.}
}
\value{
An upper-triangular matrix, in which 1 indicates linear dependence between two variates and 0 indicates independence.
}
\description{
Probability of linear dependence between two groups of time series observations using a LSSA-LFI model selection (see \code{\link[arkhaia]{LSSA_LFI_candidates}})), given a list of at least three time series. Confounding variate is selected from the remaining time series in the list.
}
