% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{model_select}
\alias{model_select}
\alias{model_select.LSSA_LFI_AIC}
\title{Model Selection of LSSA-LFI Candidates}
\usage{
model_select(x)

\method{model_select}{LSSA_LFI_AIC}(x)
}
\arguments{
\item{x}{An \code{LSSA_LFI_AIC} object.}
}
\value{
The index of the grouping which contains the lowest AIC score.
}
\description{
For an \code{LSSA_LFI_AIC} object see \code{\link[arkhaia]{LSSA_LFI_candidates}}), returns the index of the candidate model. If \code{sets} is \code{NULL} in the \code{LSSA_LFI_candidates} function, an index of [1] refers to the model of a pooled (joint) grouping, [2] refers to the model of discrete, separate groupings.
}
