% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored.R
\name{pois_rcens}
\alias{pois_rcens}
\alias{pois_rcens.matrix}
\alias{pois_rcens.data.frame}
\alias{pois_rcens.xtabs}
\alias{pois_rcens.table}
\title{Least Squares Fit of Poisson Distribution for Random Right-Censored Data}
\usage{
pois_rcens(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)

\method{pois_rcens}{matrix}(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)

\method{pois_rcens}{data.frame}(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)

\method{pois_rcens}{xtabs}(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)

\method{pois_rcens}{table}(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)
}
\arguments{
\item{x}{A matrix of cross-tabulated counts.}

\item{lambda_grid}{The resolutation at which to sample for the rate parameter. Default is \code{seq(0.1, 100, by = 0.01)}.}

\item{omit_zero}{Whether to omit zeros. Default is \code{TRUE}.}
}
\value{
A vector of the rate parameters for each column.
}
\description{
For a matrix of cross-tabulated counts of observations which constitute a minimum threshold, this function esimates the rate parameter column-wise, either retaining or omitting zeros, by a least-squares approach.
}
\examples{
x1 <- c(1,2,2,5,7,0,0)
x2 <- c(9,2,5,15,7,90,0)
x <- matrix(c(x1,x2), ncol = 2)

pois_rcens(x)
pois_rcens(x, omit_zero = FALSE)

}
