% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{repartition}
\alias{repartition}
\alias{repartition.list}
\title{Repartition (Group) Datasets}
\usage{
repartition(x, set = NULL)

\method{repartition}{list}(x, set = NULL)
}
\arguments{
\item{x}{A \code{list} containing the time series, each of which should be a matrix or data frame with time index in the first column and value in the second.}

\item{set}{A \code{list} of vectors containing the indices of the sets in \code{x}, according to which to pool the datasets into a new list.}
}
\value{
A \code{list} containing the data according to the partition structure described by \code{set}.
}
\description{
Given a list of data frames (or matrices), labeled \code{x}, and another list which contains indices of \code{x} as vectors, labeled \code{set}, returns a new list which pools (row-binds) the data frames together  according to the grouprings in \code{set}.
}
