% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{trace_pair}
\alias{trace_pair}
\alias{trace_pair.matrix}
\alias{trace_pair.data.frame}
\alias{trace_pair.xtabs}
\alias{trace_pair.table}
\title{Trace Coefficient Pairwise between Columns}
\usage{
trace_pair(x)

\method{trace_pair}{matrix}(x)

\method{trace_pair}{data.frame}(x)

\method{trace_pair}{xtabs}(x)

\method{trace_pair}{table}(x)
}
\arguments{
\item{x}{A matrix or data frame with contexts along columns and types along rows.}
}
\value{
A matrix giving the trace between all columns of the input.
}
\description{
For a contingency table with columns representing contexts and rows representing types, computes the trace of a 2 x 2 contingency table of the presence-absences across each pair of columns.
}
\examples{
x1 <- c(2, 0, 0, 11, 5, 0, 2, 0, 4)
x2 <- c(1, 1, 0, 23, 3, 3, 0, 0, 0)
x3 <- c(1, 0, 0, 0, 10, 0, 4, 0, 1)

x <- data.frame(S1 = x1, S2 = x2, S3 = x3)
rownames(x) <- LETTERS[1:nrow(x)]

trace_pair(x)

}
