% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{trim_epoch}
\alias{trim_epoch}
\alias{trim_epoch.matrix}
\alias{trim_epoch.data.frame}
\title{Trim to Epoch}
\usage{
trim_epoch(x, epoch = NULL)

\method{trim_epoch}{matrix}(x, epoch = NULL)

\method{trim_epoch}{data.frame}(x, epoch = NULL)
}
\arguments{
\item{x}{A data frame or matrix containing time-indexed data, with time index in the first column and value in the second.}

\item{epoch}{A numeric vector given the start and end time indices of the epoch.}
}
\value{
A data frame containing only those observations with time indices within the epoch.
}
\description{
For a data frame in which the first column contains a time index and the second colum observations, trim the data frame to include observations only with a given epoch (time period).
}
