% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored.R
\name{trunc_pois}
\alias{trunc_pois}
\alias{trunc_pois.matrix}
\alias{trunc_pois.data.frame}
\alias{trunc_pois.xtabs}
\alias{trunc_pois.table}
\title{Resampled Contingency Table via a Truncated Poisson for Random Right-Censored Data}
\usage{
trunc_pois(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)

\method{trunc_pois}{matrix}(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)

\method{trunc_pois}{data.frame}(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)

\method{trunc_pois}{xtabs}(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)

\method{trunc_pois}{table}(x, lambda_grid = seq(0.01, 100, by = 0.01), omit_zero = TRUE)
}
\arguments{
\item{x}{A matrix of cross-tabulated counts.}

\item{lambda_grid}{The resolutation at which to sample for the rate parameter. Default is \code{seq(0.1, 100, by = 0.01)}.}

\item{omit_zero}{Whether to omit zeros. Default is \code{TRUE}.}
}
\value{
A contingency table \eqn{Y} of the same size as \eqn{X}, with \eqn{y_{ij}} drawn according to a truncated Poisson distribution that ensures \eqn{y_{ij} \geq x_{ij}}.
}
\description{
For a matrix of cross-tabulated counts of observations which constitute a minimum threshold, returns a contingency table whose counts are sampled according to a truncated Poisson distribution, whose rate parameter is determined column-wise (see \code{\link[arkhaia]{pois_rcens}}).
}
\examples{
x1 <- c(1,2,2,5,7,0,0)
x2 <- c(9,2,5,15,7,90,0)
x <- matrix(c(x1,x2), ncol = 2)

pois_rcens(x)
pois_rcens(x, omit_zero = FALSE)

}
