% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{VB}
\alias{VB}
\alias{VB.matrix}
\alias{VB.data.frame}
\alias{VB.xtabs}
\alias{VB.table}
\title{Bias-Corrected Cramer's V}
\usage{
VB(x, lambda = 2/3)

\method{VB}{matrix}(x, lambda = 2/3)

\method{VB}{data.frame}(x, lambda = 2/3)

\method{VB}{xtabs}(x, lambda = 2/3)

\method{VB}{table}(x, lambda = 2/3)
}
\arguments{
\item{x}{A matrix or data frame of cross-tabulated counts.}

\item{lambda}{Parameter of the Cressie-Read power-divergence statistic. Default is the recommended value of 2/3.}
}
\value{
Bergsma's bias-corrected etimate of Cramer's \eqn{V}.
}
\description{
For a matrix of cross-tabulated counts of observations, estimates Cramer's V using Bergsma's bias correction \insertCite{bergsma_bias-correction_2013}{arkhaia}, using the Cressie-Read power divergence statistic (see \code{\link[arkhaia]{CR}}).
}
\examples{
x1 <- c(2, 0, 10, 11, 5)
x2 <- c(1, 1, 17, 23, 3)
x3 <- c(0, 0, 2, 81, 11)

x <- matrix(c(x1, x2, x3), ncol = 3)

VB_pair(x)
VB(x, lambda = 1)

}
\references{
\insertAllCited{}
}
