% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{VB_pair}
\alias{VB_pair}
\alias{VB_pair.matrix}
\alias{VB_pair.data.frame}
\alias{VB_pair.xtabs}
\alias{VB_pair.table}
\title{Bias-Corrected Cramer's V Pairwise between Columns}
\usage{
VB_pair(x, lambda = 2/3)

\method{VB_pair}{matrix}(x, lambda = 2/3)

\method{VB_pair}{data.frame}(x, lambda = 2/3)

\method{VB_pair}{xtabs}(x, lambda = 2/3)

\method{VB_pair}{table}(x, lambda = 2/3)
}
\arguments{
\item{x}{A matrix or data frame of cross-tabulated counts.}

\item{lambda}{Parameter of the Cressie-Read power-divergence statistic. Default is the recommended value of 2/3.}
}
\value{
A matrix of Bergsma's bias-corrected etimate of Cramer's \eqn{V}, pairwise between columns of the input matrix, of class \code{effect_sizes}.
}
\description{
For a matrix or data frame of cross-tabulated counts of observations, estimates Cramer's V using Bergsma's bias correction \insertCite{bergsma_bias-correction_2013}{arkhaia} by subsetting the table by pairs of columns (see \code{\link[arkhaia]{VB}}). In subsetting, zero row/columns are automatically removed from the subset matrix.
}
\examples{
x1 <- c(2, 0, 10, 11, 5)
x2 <- c(1, 1, 17, 23, 3)
x3 <- c(0, 0, 2, 81, 11)

x <- matrix(c(x1, x2, x3), ncol = 3)

VB_pair(x)
VB_pair(x, lambda = 1)

}
\references{
\insertAllCited{}
}
