% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored.R
\name{VB_trunc_pois}
\alias{VB_trunc_pois}
\alias{VB_trunc_pois.matrix}
\alias{VB_trunc_pois.data.frame}
\alias{VB_trunc_pois.xtabs}
\alias{VB_trunc_pois.table}
\title{Bias-Corrected Cramer's V for Random Right-Censored Data}
\usage{
VB_trunc_pois(
  x,
  lambda = 2/3,
  lambda_grid = seq(0.01, 100, by = 0.01),
  omit_zero = TRUE,
  n_iter = 10^5
)

\method{VB_trunc_pois}{matrix}(
  x,
  lambda = 2/3,
  lambda_grid = seq(0.01, 100, by = 0.01),
  omit_zero = TRUE,
  n_iter = 10^5
)

\method{VB_trunc_pois}{data.frame}(
  x,
  lambda = 2/3,
  lambda_grid = seq(0.01, 100, by = 0.01),
  omit_zero = TRUE,
  n_iter = 10^5
)

\method{VB_trunc_pois}{xtabs}(
  x,
  lambda = 2/3,
  lambda_grid = seq(0.01, 100, by = 0.01),
  omit_zero = TRUE,
  n_iter = 10^5
)

\method{VB_trunc_pois}{table}(
  x,
  lambda = 2/3,
  lambda_grid = seq(0.01, 100, by = 0.01),
  omit_zero = TRUE,
  n_iter = 10^5
)
}
\arguments{
\item{x}{A matrix of cross-tabulated counts.}

\item{lambda}{The value of lambda (default is 2/3) for the Cressie-Read power divergence statistic used to estimate the bias-corrected Cramer's \eqn{V}. Default is \code{2/3}.}

\item{lambda_grid}{The resolutation at which to sample for the rate parameter. Default is \code{seq(0.1, 100, by = 0.01)}.}

\item{omit_zero}{Whether to omit zeros. Default is \code{TRUE}.}

\item{n_iter}{Number of samples of \eqn{V} to take. Default is \code{10^5}.}
}
\value{
A contingency table \eqn{Y} of the same size as \eqn{X}, with \eqn{y_{ij}} drawn according to a truncated Poisson distribution, \eqn{y_{ij} \geq x_{ij}}.
}
\description{
Given a matrix of cross-tabulated counts of observations which constitute a minimum threshold, returns the distribution of bias-corrected Cramer's \eqn{V} by resampling from a truncated Poisson distribution, with rates determined column-wise (see \code{\link[arkhaia]{VB}}, \code{\link[arkhaia]{trunc_pois}}).
}
\examples{
x1 <- c(1,2,2,5,7,0,0)
x2 <- c(9,2,5,15,7,90,0)
x <- matrix(c(x1,x2), ncol = 2)

VB_trunc_pois(x, n_iter = 10^2)
VB_trunc_pois(x, omit_zero = FALSE, n_iter = 10^2)

}
