% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{log_OR_pair}
\alias{log_OR_pair}
\alias{log_OR_pair.matrix}
\alias{log_OR_pair.data.frame}
\alias{log_OR_pair.xtabs}
\alias{log_OR_pair.table}
\title{Log Odds Ratio with Haldane-Anscombe's Correction Pairwise between Columns}
\usage{
log_OR_pair(x)

\method{log_OR_pair}{matrix}(x)

\method{log_OR_pair}{data.frame}(x)

\method{log_OR_pair}{xtabs}(x)

\method{log_OR_pair}{table}(x)
}
\arguments{
\item{x}{A matrix or data frame with contexts along columns and types along rows.}
}
\value{
A matrix giving the log odds ratio between all columns of the input.
}
\description{
For a contingency table with columns representing contexts and rows representing types, computes the log odds ratio of a 2 x 2 contingency table of the presence-absences across each pair of columns. The addition of 0.5 to cells of the contingency table is performed \insertCite{anscombe_estimating_1956,haldane_estimation_1956}{arkhaia}.
}
\examples{
x1 <- c(2, 0, 0, 11, 5, 0, 2, 0, 4)
x2 <- c(1, 1, 0, 23, 3, 3, 0, 0, 0)
x3 <- c(1, 0, 0, 0, 10, 0, 4, 0, 1)

x <- data.frame(S1 = x1, S2 = x2, S3 = x3)
rownames(x) <- LETTERS[1:nrow(x)]

log_OR_pair(x)

}
