% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{trace_pair}
\alias{trace_pair}
\alias{trace_pair.matrix}
\title{Pairwise Trace Coefficient}
\usage{
trace_pair(x)

\method{trace_pair}{matrix}(x)
}
\arguments{
\item{x}{A data frame in which the first column is labeled \code{Context} and the second column is labeled \code{Type}, with each row containing a type present in that context.}
}
\value{
A matrix of the pairwise trace giving the overlap in types between contexts.
}
\description{
Selecting all pairs of columns (contexts) of a contingency table, computes the trace of a 2 x 2 contingency table that indicates the number of matching types (rows) that are present.
}
