% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{VB}
\alias{VB}
\alias{VB.matrix}
\title{Bias-Corrected Cramer's V}
\usage{
VB(x, method = "CR", lambda = 2/3, simulate = TRUE)

\method{VB}{matrix}(x, method = "CR", lambda = 2/3, simulate = TRUE)
}
\arguments{
\item{x}{A matrix of cross-tabulated counts.}

\item{method}{Method for estimating chi squared. Default is \code{CR}, the Cressie-Read method (see \code{\link[arkhaia]{CR}})). Other options include \code{Pearson}, which uses R's base \code{\link[stats]{chisq.test}}) function, and \code{G}, which computes the likelihood-ratio \eqn{G} statistic, which is the case for the Cressie-Read statistic as \eqn{\lambda} approaches 0.}

\item{lambda}{Only applied to the method \code{CR}. Default is the recommended value of 2/3.}

\item{simulate}{Only applied ot the method \code{Pearson}.  Whether to simulate \eqn{p} values in the \code{\link[stats]{chisq.test}}) function. Default is \code{TRUE}.
#' @examples 
x1 <- c(2, 0, 10, 11, 5)
x2 <- c(1, 1, 17, 23, 3)
x3 <- c(0, 0, 2, 81, 11)

x <- matrix(c(x1, x2, x3), ncol = 3)

VB(x)
VB(x, lambda = 1)}
}
\value{
Bergsma's bias-corrected etimate of Cramer's \eqn{V}.
}
\description{
For a matrix of cross-tabulated counts of observations, estimates Cramer's V using Bergsma's bias correction \insertCite{bergsma_bias-correction_2013}{arkhaia}.
}
\references{
\insertAllCited{}
}
