% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{VB_LOO_type}
\alias{VB_LOO_type}
\alias{VB_LOO_type.data.frame}
\title{Leave-One-Out Routine for Pairwise Effect Sizes of Archaeological Contexts}
\usage{
VB_LOO_type(x, lambda = 2/3)

\method{VB_LOO_type}{data.frame}(x, lambda = 2/3)
}
\arguments{
\item{x}{A data frame in which the first column is labeled \code{Context} and the second column is labeled \code{Type}, with each row containing a type present in that context.}

\item{lambda}{The Cressie-Read power-divergence statistic parameter. Default is the recommended value of 2/3.}
}
\value{
A three-dimensional array of the pairwise context-by-context effect sizes given the ommission of a type, for all types given in the input data frame.
}
\description{
To evaluate the stability of estimates of effect sizes between archaeological contexts in light of the inclusion or exclusion of a given type, this routine computes bias-corrected Cramer's V \code{\link[arkhaia]{VB}}) omitting a type on each iteration.
}
